Rename project to project-card
