+++ src/components/project-card/project-card.scss
