+++ src/components/project-overview/project-overview.tsx
@@ -1,8 +1,8 @@
-import Project from '@/components/project/project';
+import { Project as ProjectInterface } from '@/app/dataModels/project';
+import Project from '@/components/project-card/project-card';
 import { getLocalProjectData } from '@/utils/get-local-project-data/get-local-project-data';
 import { Box, Card } from '@mui/material';
 import { Fragment } from 'react';
-import { Project as ProjectInterface } from '@/app/dataModels/project';
 import './project-overview.scss';
 
 export default async function ProjectOverview() {
